Refactor: fix barrel imports circular dependencies (#10058)
* fix few barrel dependencies cycles

* fix few barrel dependencies cycles

* Refactor: fix editor barrel imports (#10078)

* fix editor barrel imports

* Refactor: Break 4 circular dependencies (#10079)

* break circular deps: file extraction, barrel bypass, class co-location

* break circular deps: file extraction, barrel bypass, class co-location

* break circular deps: file extraction, barrel bypass, class co-location

* break RGRoupAttachmentPoint circular dep

* break RGRoupAttachmentPoint circular dep
