+++ packages/ketcher-core/src/application/editor/internal.ts
@@ -1,4 +1,3 @@
-export * from './Editor';
 export * from './EditorHistory';
 export * from './shared/coordinates';
 export * from './editor.types';
