+++ packages/ketcher-core/src/application/editor/operations/rgroupAttachmentPoint/index.ts
@@ -1,2 +1,7 @@
-export * from './RGroupAttachmentPointAdd';
-export * from './RGroupAttachmentPointRemove';
+import { RGroupAttachmentPointAdd } from './RGroupAttachmentPointAdd';
+import { RGroupAttachmentPointRemove } from './RGroupAttachmentPointRemove';
+
+RGroupAttachmentPointAdd.InverseConstructor = RGroupAttachmentPointRemove;
+RGroupAttachmentPointRemove.InverseConstructor = RGroupAttachmentPointAdd;
+
+export { RGroupAttachmentPointAdd, RGroupAttachmentPointRemove };
