+++ packages/ketcher-core/src/application/render/renderers/TransientView/RotationView.ts
@@ -17,7 +17,7 @@
 import { TransientView } from 'application/render/renderers/TransientView/TransientView';
 import { D3SvgElementSelection } from 'application/render/types';
 import { Vec2 } from 'domain/entities';
-import { Coordinates } from 'application/editor';
+import { Coordinates } from 'application/editor/shared/coordinates';
 
 export type RotationViewParams = {
   center: Vec2;
